Move to an undocumented job API to support newer Connect versions.
The filesystem format of the logs was changed recently and bundle IDs
are no longer visible. Instead, use an undocumented API to map job IDs
to content GUIDs.

PIDs are no longer easily accessible, so they have been removed.

On the bright side, this avoids the need to read metadata off the
filesystem.